refact : add model folder and move classes
