+++ src/test/java/baseball/model/ComputerTest.java
@@ -1,7 +1,8 @@
-package baseball;
+package baseball.model;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import baseball.model.Computer;
 import baseball.validate.NumbersValidate;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
